requirements.txt
absl-py==2.4.0
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
deepspeed==0.18.9
diffusers==0.37.1
einops==0.8.2
filelock==3.28.0
flash_attn==2.8.3
fsspec==2026.3.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpi4py==4.1.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==26.1
peft==0.19.1
pillow==12.2.0
pip==26.0.1
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
setuptools==82.0.1
shellingham==1.5.4
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
transformers==5.5.4
triton==3.4.0
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
Werkzeug==3.1.8
wheel==0.46.3
zipp==3.23.1
